Split out and simplify conflict types.
