+++ src/commands/scaffoldDaprTasks.ts
@@ -13,7 +13,7 @@ import { UserInput, WizardStep } from '../services/userInput';
 import { IActionContext, TelemetryProperties } from 'vscode-azureextensionui';
 import { TemplateScaffolder } from '../scaffolding/templateScaffolder';
 import { Scaffolder } from '../scaffolding/scaffolder';
-import { TaskConflictHandler } from '../scaffolding/taskScaffolder';
+import { ConflictHandler } from '../scaffolding/conflicts';
 
 interface ScaffoldTelemetryProperties extends TelemetryProperties {
     configurationType: string;
@@ -92,7 +92,7 @@ function getDefaultPort(configuration: vscode.DebugConfiguration | undefined): n
     return DefaultPort;
 }
 
-function* range(start: number = 0) {
+function* range(start = 0) {
     while (true) {
         yield start++;
     }
@@ -189,17 +189,17 @@ export async function scaffoldDaprTasks(context: IActionContext, scaffolder: Sca
 
     telemetryProperties.configurationType = result.configuration.type;
 
-    const onTaskConflict: TaskConflictHandler =
-        async (task, isUnique) => {
+    const onTaskConflict: ConflictHandler =
+        async (label, isUnique) => {
             const overwrite: vscode.MessageItem = { title: localize('commands.scaffoldDaprTasks.overwriteTask', 'Overwrite') };
             const newTask: vscode.MessageItem = { title: localize('commands.scaffoldDaprTasks.createTask', 'Create task') };
 
-            const result = await ui.showWarningMessage(localize('commands.scaffoldDaprTasks.taskExists', 'The task \'{0}\' already exists. Do you want to overwrite it or create a new task?', task.label), overwrite, newTask);
+            const result = await ui.showWarningMessage(localize('commands.scaffoldDaprTasks.taskExists', 'The task \'{0}\' already exists. Do you want to overwrite it or create a new task?', label), overwrite, newTask);
 
             if (result === overwrite) {
                 return { 'type': 'overwrite' };
             } else {
-                const generator = nameGenerator(task.label ?? '', range());
+                const generator = nameGenerator(`${label ?? ''}-`, range(1));
                 let name = generator.next();
 
                 while (!name.done && !await isUnique(name.value)) {
@@ -210,7 +210,7 @@ export async function scaffoldDaprTasks(context: IActionContext, scaffolder: Sca
                     throw new Error(localize('commands.scaffoldDaprTasks.renameError', 'Unable to generate a unique task.'));
                 }
 
-                return { 'type': 'rename', value: name.value };
+                return { 'type': 'rename', name: name.value };
             }
         };
