+++ src/scaffolding/conflicts.ts
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export interface OverwriteResult {
+    type: 'overwrite';
+}
+
+export interface RenameResult {
+    type: 'rename';
+    name: string;
+}
+
+export interface SkipResult {
+    type: 'skip';
+}
+
+export type ConflictResult = OverwriteResult | RenameResult | SkipResult;
+
+export type ConflictHandler = (name: string, isUnique: (name: string) => Promise<boolean>) => Promise<ConflictResult>;
