+++ src/scaffolding/scaffolder.ts
@@ -1,10 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
 import scaffoldConfiguration, { DebugConfiguration } from './configurationScaffolder';
-import scaffoldTask, { TaskContentFactory, TaskConflictHandler } from './taskScaffolder';
+import scaffoldTask, { TaskContentFactory } from './taskScaffolder';
+import { ConflictHandler } from './conflicts';
 
 export interface Scaffolder {
     scaffoldConfiguration(name: string, contentFactory: (name: string) => DebugConfiguration, onConflict: (configuration: DebugConfiguration) => Promise<boolean>): Promise<string | undefined>;
     scaffoldFile(fileName: string, contentFactory: (fileName: string) => string, onConflict: (fileName: string) => Promise<boolean>): Promise<string | undefined>;
-    scaffoldTask(label: string, contentFactory: TaskContentFactory, onConflict: TaskConflictHandler): Promise<string | undefined>;
+    scaffoldTask(label: string, contentFactory: TaskContentFactory, onConflict: ConflictHandler): Promise<string | undefined>;
 }
 
 export default class LocalScaffolder implements Scaffolder {
@@ -22,7 +26,7 @@ export default class LocalScaffolder implements Scaffolder {
         return Promise.resolve(undefined);
     }
 
-    scaffoldTask(label: string, contentFactory: TaskContentFactory, onConflict: TaskConflictHandler): Promise<string | undefined> {
+    scaffoldTask(label: string, contentFactory: TaskContentFactory, onConflict: ConflictHandler): Promise<string | undefined> {
         return scaffoldTask(label, contentFactory, onConflict);
     }
 }
